修改可以调节反馈的阈值

+++ app/src/main/jni/Dehazor.cpp
@@ -36,7 +36,7 @@ JNIEXPORT void JNICALL Java_com_willhua_opencvstudy_OpenCVMethod_fastDehazor(JNI
     LOG("fast   dehazor  end");
 }
 
-JNIEXPORT void JNICALL Java_com_willhua_opencvstudy_OpenCVMethod_fastDehazorCV(JNIEnv *env, jclass cls, jobject bitmap, int width, int height, int radius, float p, float s) {
+JNIEXPORT void JNICALL Java_com_willhua_opencvstudy_OpenCVMethod_fastDehazorCV(JNIEnv *env, jclass cls, jobject bitmap, int width, int height, int radius, float p, int s) {
     LOG("fast  dehazor  begin");
     char * data;
